alerts_service/requirements.txt
kafka-python==2.0.3
requests
SQLAlchemy
psycopg2-binary
dotenv
